# Conflicts:
#	app/build.gradle
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/example/studybuddies/MainMenuUI.java
#	app/src/main/java/com/example/studybuddies/MainMenuViewHolder.java
#	app/src/main/java/com/example/studybuddies/Post.java
#	app/src/main/java/com/example/studybuddies/createPostUI.java
#	app/src/main/res/layout/activity_main.xml
#	app/src/main/res/layout/create_post.xml
#	app/src/main/res/layout/main_menu_ui.xml
#	app/src/main/res/layout/single_view_mainmenu_layout.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/studybuddies/Post.java
@@ -8,13 +8,13 @@ public class Post {
     String postTitle;
     String postDescription;
     String subject;
-    double ratings;
+    int ratings;
 
     public Post(){
 
     }
 
-    public Post(String postID, String postTitle, String postDescription, String subject, double ratings){
+    public Post(String postID, String postTitle, String postDescription, String subject, int ratings){
         this.postID = postID;
         this.postTitle = postTitle;
         this.postDescription = postDescription;
@@ -38,7 +38,7 @@ public class Post {
         return subject;
     }
 
-    public double getRatings(){
+    public int getRatings(){
         return ratings;
     }
 }
